Moved Registration to component folder
