+++ src/App.css
@@ -3,6 +3,11 @@
     font-family: 'Roboto', sans-serif;
 */
 
+
+
+
+/*  Registration Component CSS Below */
+
 .reg-title{
     font-family: 'Ubuntu', sans-serif;
     font-size: 3em;
