+++ src/App.js
@@ -1,7 +1,7 @@
 import React, {useState, useEffect} from 'react';
 import Game from './components/Game'
 import './App.css';
-import RegF from './Registration'
+import RegF from './components/Registration'
 
 function App() {
   return (
